fix: FAB to IconButton, HostButton target is absolute, refactoring
