+++ src/components/HostButton/index.tsx
@@ -1,20 +1,18 @@
 import * as React from "react";
-import { useSocket } from "../App";
-import LogoutIcon from "@mui/icons-material/Logout";
+import GamepadIcon from "@mui/icons-material/Gamepad";
 import Fab from "@mui/material/Fab";
 import WarningIcon from "@mui/icons-material/Warning";
 import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
 import { ClickAwayListener } from "@mui/material";
-import { leaveConfirmBox, leaveFAB } from "./styles";
+import { Link as RouterLink } from "react-router-dom";
+import {
+    hostConfirmBox,
+    hostFAB,
+} from "./styles";
 
-function Leave() {
-    const { socket } = useSocket();
-
-    const handleLeave = () => {
-        socket.emit("leave");
-    };
+function HostButton() {
     const [ open, setOpen ] = React.useState(false);
     const handleClick = () => {
         setOpen((prev) => !prev);
@@ -23,37 +21,43 @@ function Leave() {
         setOpen(false);
     };
 
-
     return (
-        <React.Fragment>
+        <div className={"create-game-fab"}>
             <ClickAwayListener onClickAway={handleClickAway}>
                 <Box>
                     <Fab
                         size="small"
-                        color="secondary"
-                        aria-label="logout"
-                        sx={leaveFAB}
+                        color="primary"
+                        aria-label="new"
+                        sx={hostFAB}
                         onClick={handleClick}
                     >
-                        <LogoutIcon />
+                        <GamepadIcon />
                     </Fab>
                     {open
                         ? (
-                            <Box sx={leaveConfirmBox}>
+                            <Box sx={hostConfirmBox}>
                                 <p style={{ margin: "0 0 0.5em 0" }}><WarningIcon />
-                                    Want to leave the room?</p>
+                                    Host new game?</p>
                                 <Stack spacing={2} direction="row"
                                     style={{ justifyContent: "center" }}>
                                     <Button variant="outlined" onClick={handleClickAway}>No</Button>
-                                    <Button variant="contained" onClick={handleLeave}>Yes</Button>
+                                    <Button
+                                        variant="contained"
+                                        component={RouterLink}
+                                        to="/host"
+                                        onClick={handleClickAway}
+                                    >
+                                        Yes
+                                    </Button>
                                 </Stack>
                             </Box>
                         )
                         : null}
                 </Box>
             </ClickAwayListener>
-        </React.Fragment>
+        </div>
     );
 }
 
-export default Leave;
+export default HostButton;
