+++ src/components/LeaveButton/index.tsx
@@ -1,18 +1,20 @@
 import * as React from "react";
-import GamepadIcon from "@mui/icons-material/Gamepad";
+import { useSocket } from "../App";
+import LogoutIcon from "@mui/icons-material/Logout";
 import Fab from "@mui/material/Fab";
 import WarningIcon from "@mui/icons-material/Warning";
 import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
 import { ClickAwayListener } from "@mui/material";
-import { Link as RouterLink } from "react-router-dom";
-import {
-    hostConfirmBox,
-    hostFAB,
-} from "./styles";
+import { leaveConfirmBox, leaveFAB } from "./styles";
 
-function CreateGame() {
+function LeaveButton() {
+    const { socket } = useSocket();
+
+    const handleLeave = () => {
+        socket.emit("leave");
+    };
     const [ open, setOpen ] = React.useState(false);
     const handleClick = () => {
         setOpen((prev) => !prev);
@@ -21,43 +23,37 @@ function CreateGame() {
         setOpen(false);
     };
 
+
     return (
-        <div className={"create-game-fab"}>
+        <React.Fragment>
             <ClickAwayListener onClickAway={handleClickAway}>
                 <Box>
                     <Fab
                         size="small"
-                        color="primary"
-                        aria-label="new"
-                        sx={hostFAB}
+                        color="secondary"
+                        aria-label="logout"
+                        sx={leaveFAB}
                         onClick={handleClick}
                     >
-                        <GamepadIcon />
+                        <LogoutIcon />
                     </Fab>
                     {open
                         ? (
-                            <Box sx={hostConfirmBox}>
+                            <Box sx={leaveConfirmBox}>
                                 <p style={{ margin: "0 0 0.5em 0" }}><WarningIcon />
-                                    Host new game?</p>
+                                    Want to leave the room?</p>
                                 <Stack spacing={2} direction="row"
                                     style={{ justifyContent: "center" }}>
                                     <Button variant="outlined" onClick={handleClickAway}>No</Button>
-                                    <Button
-                                        variant="contained"
-                                        component={RouterLink}
-                                        to="host"
-                                        onClick={handleClickAway}
-                                    >
-                                        Yes
-                                    </Button>
+                                    <Button variant="contained" onClick={handleLeave}>Yes</Button>
                                 </Stack>
                             </Box>
                         )
                         : null}
                 </Box>
             </ClickAwayListener>
-        </div>
+        </React.Fragment>
     );
 }
 
-export default CreateGame;
+export default LeaveButton;
