+++ src/components/MenuButtons/index.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
-import CreateGame from "../CreateGame";
+import HostButton from "../HostButton";
 import { LibraryButton } from "../LibraryButton";
 
 export function MenuButtons(): JSX.Element {
     return <>
         <LibraryButton />
-        <CreateGame />
+        <HostButton />
     </>;
 }
